Update subplacegui.ps1 to v1.1

+++ subplacegui.ps1
@@ -1,17 +1,17 @@
 <#
 subplacegui.ps1
-by kit, version 1.0
+by kit, version 1.1
 subject to the terms of the MPL 2.0, you can get a copy at http://mozilla.org/MPL/2.0/
 #>
 
-$version  = "1.0"
+$version  = "1.1"
 $jsonPath = "$env:LOCALAPPDATA\subplace\gui\settings.json"
 Add-Type -AssemblyName PresentationFramework
 
 $xaml = @"
 <Window xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	Title="subplacegui.ps1" Height="450" Width="600">
+	Title="subplacegui" Height="450" Width="600">
 	<TabControl x:Name="MainTabControl">
 		<TabItem Header="Main">
 			<Grid>
@@ -30,8 +30,10 @@ $xaml = @"
 						<DataGridTextColumn Header="ID" Binding="{Binding id}" Width="Auto"/>
 					</DataGrid.Columns>
 				</DataGrid>
-				<!-- <StackPanel Grid.Row="2" Margin="10">
-				</StackPanel> -->
+				<StackPanel Grid.Row="2" Margin="10,0,10,10" Orientation="Horizontal">
+					<TextBlock x:Name="StatusDot" Foreground="Gray" FontFamily="Segoe UI Symbol Regular">&#x26AB;</TextBlock>
+					<TextBlock x:Name="StatusText">&#x2002;Joining process hasn't begun yet.</TextBlock>
+				</StackPanel>
 			</Grid>
 		</TabItem>
 		<TabItem Header="Settings">
@@ -44,13 +46,14 @@ $xaml = @"
 				<StackPanel Margin="0" Grid.Row="0">
 					<StackPanel Orientation="Horizontal" Margin="-5,0,0,0">
 						<Label Content="Delay:" Target="{Binding ElementName=Delay}"/>
-						<TextBox x:Name="Delay" Text="750" HorizontalAlignment="Left" Width="150" Margin="0,0,10,0" Height="18"/>
+						<TextBox x:Name="Delay" Text="750" HorizontalAlignment="Left" Width="150" Height="18"/>
 					</StackPanel>
 					<StackPanel Orientation="Horizontal" Margin="-5,0,0,8">
 						<Label Content="Loop Delay:" Target="{Binding ElementName=LoopDelay}"/>
-						<TextBox x:Name="LoopDelay" Text="300" HorizontalAlignment="Left" Width="150" Margin="0,0,10,0" Height="18"/>
+						<TextBox x:Name="LoopDelay" Text="300" HorizontalAlignment="Left" Width="150" Height="18"/>
 					</StackPanel>
-					<CheckBox x:Name="SudoMode" Content="Enable sudo mode" Margin="0,0,0,0"/>
+					<CheckBox x:Name="SudoMode" Content="Enable sudo mode" Margin="0,0,0,10"/>
+					<CheckBox x:Name="SkipPrejoining" Content="Skip pre-joining" Margin="0,0,0,10"/>
 				</StackPanel>
 					<TextBlock x:Name="SavedMessage" Text="Settings Saved" Foreground="Green" Margin="0,-10,0,10" Grid.Row="2" Visibility="Hidden"/>
 				<Button Content="Save Settings" 
@@ -61,6 +64,14 @@ $xaml = @"
 					Grid.Row="2"/>
 			</Grid>
 		</TabItem>
+		<TabItem Header="Info">
+			<StackPanel Margin="10">
+				<TextBlock Margin="0,0,0,5">subplacegui 1.1 by Kit</TextBlock>
+				<TextBlock Margin="0,0,0,5"><Hyperlink x:Name="GithubLink" NavigateUri="https://github.com/catb0x/subplace">Github</Hyperlink></TextBlock>
+				<TextBlock Margin="0,0,0,5" Foreground="Gray">stan loona :3</TextBlock>
+    				<TextBlock Margin="0,0,0,5" Foreground="Gray">rgc was here</TextBlock>
+			</StackPanel>
+		</TabItem>
 	</TabControl>
 </Window>
 "@
@@ -68,14 +79,18 @@ $xaml = @"
 $reader = [System.Xml.XmlReader]::Create([System.IO.StringReader]::new($xaml))
 $window = [Windows.Markup.XamlReader]::Load($reader)
 
-$placeIdInput		= $window.FindName("PlaceIdInput")
+$placeIdInput       = $window.FindName("PlaceIdInput")
 $setPlaceIdButton   = $window.FindName("SetPlaceIdButton")
-$placesGrid			= $window.FindName("PlacesGrid")
-$delayInput			= $window.FindName("Delay")
-$loopDelayInput	 = $window.FindName("LoopDelay")
+$placesGrid         = $window.FindName("PlacesGrid")
+$delayInput         = $window.FindName("Delay")
+$loopDelayInput     = $window.FindName("LoopDelay")
 $sudoModeCheckbox   = $window.FindName("SudoMode")
 $saveSettingsButton = $window.FindName("SaveSettingsButton")
 $savedMessageText   = $window.FindName("SavedMessage")
+$skipPrejoining     = $window.FindName("SkipPrejoining")
+$statusText         = $window.FindName("StatusText")
+$statusDot          = $window.FindName("StatusDot")
+$githubLink         = $window.FindName("GithubLink")
 
 $fadeOut = New-Object Windows.Media.Animation.DoubleAnimation
 $fadeOut.From = 1.0
@@ -88,6 +103,10 @@ if (!([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]
 } elseif (!(Get-Command "gsudo" -ErrorAction SilentlyContinue)) {
 	$sudoModeCheckbox.Visibility = [System.Windows.Visibility]::Collapsed
 }
+
+$githubLink.Add_Click({
+	Start-Process $this.NavigateUri.AbsoluteUri
+})
 
 $NumberValidationTextBox = {
 	param($sender, $e)
@@ -114,8 +133,8 @@ $setPlaceIdButton.Add_Click({
 	$placeId = $placeIdInput.Text
 	if (-not [string]::IsNullOrEmpty($placeId)) {
 		try {
-			$request	   = Invoke-RestMethod -Uri "https://apis.roblox.com/universes/v1/places/$placeId/universe"
-			$universeId	   = $request.universeId
+			$request       = Invoke-RestMethod -Uri "https://apis.roblox.com/universes/v1/places/$placeId/universe"
+			$universeId    = $request.universeId
 			$rootRequest   = Invoke-RestMethod -Uri "https://games.roblox.com/v1/games?universeIds=$universeId"
 			$script:rootId = $rootRequest.data[0].rootPlaceId
 			if ($placeId -eq $script:rootId) {
@@ -192,31 +211,35 @@ function LoadSettings {
 			Set-Variable -Name $key -Value $varHash.$key -Scope Script -Force
 		}
 	} else {
-		$script:delay	 = 750
+		$script:delay     = 750
 		$script:loopDelay = 300
 		$script:sudoMode  = $false
 		SaveSettings "delay", "loopDelay", "sudoMode", "version"
 	}
-	$script:delayInput.Text			= $delay
-	$script:loopDelayInput.Text		= $loopDelay
+	$script:delayInput.Text            = $delay
+	$script:loopDelayInput.Text        = $loopDelay
 	$script:sudoModeCheckbox.IsChecked = $sudoMode
 }
 LoadSettings
 
 function Join {
-	if ($sudoMode) {
-		gsudo --integrity medium Start-Process "roblox://experiences/start?placeId=$rootId"
-	} else {
-		Start-Process "roblox://experiences/start?placeId=$rootId"
+	$statusDot.Foreground = [System.Windows.Media.Brushes]::Red
+	$statusText.Text      = " Do not click on retry yet."
+	if (-not ($skipPrejoining.Checked)) {
+		if ($sudoMode) {
+			gsudo --integrity medium Start-Process "roblox://experiences/start?placeId=$rootId"
+		} else {
+			Start-Process "roblox://experiences/start?placeId=$rootId"
+		}
+		while (-not (WindowInForeground "Roblox")) {
+			Start-Sleep -Milliseconds $loopDelay
+		}
+		$process     = WindowInForeground "Roblox"
+		$processPath = (Get-Process -Id $process.Id).Path
+		$processName = [System.IO.Path]::GetFileNameWithoutExtension($processPath)
+		Start-Sleep -Milliseconds $delay
+		Stop-Process -Name $processName
 	}
-	while (-not (WindowInForeground "Roblox")) {
-		Start-Sleep -Milliseconds $loopDelay
-	}
-	$process	 = WindowInForeground "Roblox"
-	$processPath = (Get-Process -Id $process.Id).Path
-	$processName = [System.IO.Path]::GetFileNameWithoutExtension($processPath)
-	Start-Sleep -Milliseconds $delay
-	Stop-Process -Name $processName
 	if ($sudoMode) {
 		gsudo --integrity medium Start-Process "roblox://experiences/start?placeId=$subplace"
 	} else {
@@ -228,6 +251,16 @@ function Join {
 	BlockInternet $processPath
 	Start-Sleep -Milliseconds $delay
 	UnblockInternet $processPath
+	$statusDot.Foreground = [System.Windows.Media.Brushes]::Green
+	$statusText.Text = " You can click on retry now."
+	$script:timer = New-Object System.Windows.Threading.DispatcherTimer
+	$timer.Interval = [TimeSpan]::FromSeconds(5)
+	$timer.Add_Tick({
+		$statusDot.Foreground = [System.Windows.Media.Brushes]::Gray
+		$statusText.Text = " Joining process hasn't begun yet."
+		$timer.Stop()
+	})
+$timer.Start()
 }
 
 function WindowInForeground {
